--- src/main/java/minijava/lexer/SimpleLexer.java
@@ -1,454 +1,0 @@
-package minijava.lexer;
-
-import static minijava.token.Terminal.*;
-
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Maps;
-import java.io.ByteArrayInputStream;
-import java.util.ArrayList;
-import java.util.EnumSet;
-import java.util.List;
-import minijava.MJError;
-import minijava.token.Position;
-import minijava.token.Terminal;
-import minijava.token.Token;
-
-/** SLL(1) parser style lexer implementation. */
-public class SimpleLexer implements Lexer {
-
-  private LexerInput input;
-  private Token current = null;
-  private Position position;
-  private List<Token> lookAheadBuffer = new ArrayList<>();
-  private int numberOfEOFs = 0;
-
-  static final ImmutableMap<String, Terminal> keywords =
-      Maps.uniqueIndex(
-          EnumSet.of(
-              BOOLEAN, CLASS, ELSE, FALSE, IF, INT, NEW, NULL, PUBLIC, RETURN, STATIC, THIS, TRUE,
-              VOID, WHILE),
-          token -> token.getDescription());
-
-  static final ImmutableSet<String> reservedIdentifiers =
-      ImmutableSet.of(
-          "abstract",
-          "assert",
-          "break",
-          "byte",
-          "case",
-          "catch",
-          "char",
-          "const",
-          "continue",
-          "default",
-          "double",
-          "do",
-          "enum",
-          "extends",
-          "finally",
-          "final",
-          "float",
-          "for",
-          "goto",
-          "implements",
-          "import",
-          "instanceof",
-          "interface",
-          "long",
-          "native",
-          "package",
-          "private",
-          "protected",
-          "short",
-          "strictfp",
-          "super",
-          "switch",
-          "synchronized",
-          "throws",
-          "throw",
-          "transient",
-          "try",
-          "volatile");
-
-  public SimpleLexer(LexerInput input) {
-    this.input = input;
-  }
-
-  private Token parseNextToken() {
-    if (input.current() > 127) {
-      throw new LexerError(
-          input.getCurrentPosition(), String.format("Invalid char number %d", input.current()));
-    }
-    if (input.current() <= 0) {
-      position = input.getCurrentPosition();
-      return createToken(EOF, "");
-    }
-    omitWS();
-    if (input.current() <= 0) {
-      position = input.getCurrentPosition();
-      return createToken(EOF, "");
-    }
-    position = input.getCurrentPosition();
-    byte cur = (byte) input.current();
-    if (Character.isDigit(cur)) {
-      return parseInt();
-    }
-    if (isAlphabet(cur) || cur == '_') {
-      return parseKeywordOrIdentifier();
-    }
-    switch (cur) {
-      case '+':
-        input.next();
-        return parsePlus();
-      case '>':
-        input.next();
-        return parseGreater();
-      case '(':
-        input.next();
-        return createToken(LPAREN, "(");
-      case ')':
-        input.next();
-        return createToken(RPAREN, ")");
-      case '?':
-        input.next();
-        return createToken(QUESTION_MARK, "?");
-      case ';':
-        input.next();
-        return createToken(SEMICOLON, ";");
-      case '[':
-        input.next();
-        return createToken(LBRACKET, "[");
-      case ']':
-        input.next();
-        return createToken(RBRACKET, "]");
-      case '/':
-        input.next();
-        return parseSlash();
-      case '-':
-        input.next();
-        return parseMinus();
-      case '{':
-        input.next();
-        return createToken(LCURLY, "{");
-      case '}':
-        input.next();
-        return createToken(RCURLY, "}");
-      case ':':
-        input.next();
-        return createToken(COLON, ":");
-      case ',':
-        input.next();
-        return createToken(COMMA, ",");
-      case '%':
-        input.next();
-        return parseModulo();
-      case '.':
-        input.next();
-        return createToken(DOT, ".");
-      case '<':
-        input.next();
-        return parseLower();
-      case '=':
-        input.next();
-        return parseEqual();
-      case '!':
-        input.next();
-        return parseInvert();
-      case '&':
-        input.next();
-        return parseAnd();
-      case '~':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "~");
-      case '*':
-        input.next();
-        return parseStar();
-      case '|':
-        input.next();
-        return parsePipe();
-      case '^':
-        input.next();
-        return parseCaret();
-      default:
-        throw createError();
-    }
-  }
-
-  private MJError createError() {
-    return new LexerError(
-        input.getCurrentPosition(),
-        String.format("Unexpected character '%s'(%d)", input.current(), input.current()));
-  }
-
-  private void omitWS() {
-    while (input.isCurrentChar(' ', '\n', '\r', '\t')) {
-      input.next();
-    }
-  }
-
-  private Token parseInvert() {
-    switch (input.current()) {
-      case '=':
-        input.next();
-        return createToken(UNEQUALS, "!=");
-      default:
-        return createToken(INVERT, "!");
-    }
-  }
-
-  private Token parseCaret() {
-    switch (input.current()) {
-      case '=':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "^=");
-      default:
-        return createToken(RESERVED_OPERATORS, "^");
-    }
-  }
-
-  private Token parseModulo() {
-    switch (input.current()) {
-      case '=':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "%=");
-      default:
-        return createToken(MODULO, "%");
-    }
-  }
-
-  private Token parseInt() {
-    StringBuilder builder = new StringBuilder();
-    builder.appendCodePoint(input.current());
-    if (input.current() == '0') {
-      input.next();
-    } else {
-      // range [1-9]
-      byte cur = input.next();
-      while (cur >= '0' && cur <= '9') {
-        builder.appendCodePoint(cur);
-        cur = input.next();
-      }
-    }
-    return createToken(INTEGER_LITERAL, builder.toString());
-  }
-
-  private Token parsePlus() {
-    switch (input.current()) {
-      case '+':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "++");
-      case '=':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "+=");
-      default:
-        return createToken(PLUS, "+");
-    }
-  }
-
-  private Token parseGreater() {
-    switch (input.current()) {
-      case '>':
-        input.next();
-        switch (input.current()) {
-          case '>':
-            input.next();
-            switch (input.current()) {
-              case '=':
-                input.next();
-                return createToken(RESERVED_OPERATORS, ">>>=");
-              default:
-                return createToken(RESERVED_OPERATORS, ">>>");
-            }
-          case '=':
-            input.next();
-            return createToken(RESERVED_OPERATORS, ">>=");
-          default:
-            return createToken(RESERVED_OPERATORS, ">>");
-        }
-      case '=':
-        input.next();
-        return createToken(GREATER_EQUALS, ">=");
-      default:
-        return createToken(GREATER, ">");
-    }
-  }
-
-  private Token parseSlash() {
-    switch (input.current()) {
-      case '*':
-        input.next();
-        return parseCommentRest();
-      case '=':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "/=");
-      default:
-        return createToken(DIVIDE, "/");
-    }
-  }
-
-  private Token parseCommentRest() {
-    while (true) {
-      byte cur = input.current();
-      byte next = input.next();
-      if (cur < 0 || next < 0) {
-        throw createError();
-      }
-      if (cur == '*' && next == '/') {
-        input.next();
-        return createToken(COMMENT, "");
-      }
-    }
-  }
-
-  private Token parseMinus() {
-    switch (input.current()) {
-      case '=':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "-=");
-      case '-':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "--");
-      default:
-        return createToken(MINUS, "-");
-    }
-  }
-
-  private Token parseLower() {
-    switch (input.current()) {
-      case '<':
-        input.next();
-        switch (input.current()) {
-          case '=':
-            input.next();
-            return createToken(RESERVED_OPERATORS, "<<=");
-          default:
-            return createToken(RESERVED_OPERATORS, "<<");
-        }
-      case '=':
-        input.next();
-        return createToken(LOWER_EQUALS, "<=");
-      default:
-        return createToken(LOWER, "<");
-    }
-  }
-
-  private Token parseEqual() {
-    switch (input.current()) {
-      case '=':
-        input.next();
-        return createToken(EQUALS, "==");
-      default:
-        return createToken(EQUAL_SIGN, "=");
-    }
-  }
-
-  private Token parseAnd() {
-    switch (input.current()) {
-      case '&':
-        input.next();
-        return createToken(AND, "&&");
-      case '=':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "&=");
-      default:
-        return createToken(RESERVED_OPERATORS, "&");
-    }
-  }
-
-  private Token parseStar() {
-    switch (input.current()) {
-      case '=':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "*=");
-      default:
-        return createToken(MULTIPLY, "*");
-    }
-  }
-
-  private Token parsePipe() {
-    switch (input.current()) {
-      case '|':
-        input.next();
-        return createToken(OR, "||");
-      case '=':
-        input.next();
-        return createToken(RESERVED_OPERATORS, "|=");
-      default:
-        return createToken(RESERVED_OPERATORS, "|");
-    }
-  }
-
-  private Token parseKeywordOrIdentifier() {
-    StringBuilder builder = new StringBuilder();
-    builder.appendCodePoint(input.current());
-    byte cur = (byte) (int) input.next();
-    while (input.current() != -1 && (isAlphabet(cur) || cur == '_' || Character.isDigit(cur))) {
-      builder.appendCodePoint(cur);
-      cur = (byte) (int) input.next();
-    }
-    String word = builder.toString();
-    Terminal keywordTerminal = keywords.get(word);
-    if (keywordTerminal != null) {
-      return createToken(keywordTerminal, word);
-    }
-    if (reservedIdentifiers.contains(word)) {
-      return createToken(RESERVED_IDENTIFIER, word);
-    }
-    return createToken(IDENT, builder.toString());
-  }
-
-  private boolean isAlphabet(byte c) {
-    return (c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z');
-  }
-
-  private Token createToken(Terminal terminal, String content) {
-    return new Token(terminal, position, content);
-  }
-
-  private Token nextToken() {
-    if (current != null && current.isEOF()) {
-      return current;
-    }
-    if (lookAheadBuffer.size() > 0) {
-      current = lookAheadBuffer.get(0);
-      lookAheadBuffer.remove(0);
-    } else {
-      current = parseNextToken();
-    }
-    return current;
-  }
-
-  @Override
-  public Token current() {
-    if (current == null) {
-      current = parseNextToken();
-      if (current.isEOF()) {
-        numberOfEOFs++;
-      }
-    }
-    return current;
-  }
-
-  @Override
-  public boolean hasNext() {
-    return numberOfEOFs < 1;
-  }
-
-  @Override
-  public Token next() {
-    current = nextToken();
-    while (current.terminal == COMMENT) {
-      current = nextToken();
-    }
-    if (current.isEOF()) {
-      numberOfEOFs++;
-    }
-    return current;
-  }
-
-  public static List<Token> getAllTokens(String input) {
-    return LexerUtils.getAllTokens(
-        new SimpleLexer(new BasicLexerInput(new ByteArrayInputStream(input.getBytes()))));
-  }
-}
